Allowing game updates to be ready once init finished.
This change is because we no longer make use of the old code that tries to send in galaxy/sector data. That will eventually be replaced with a proper mod using the new plugin architecture.

Fixed a bug with logging that caused the file descriptor for our logfile to always be undefined.